feat: incremented member and project, created member_project and ProfessionalProfile, fixed some tests and adapted

+++ src/modules/project/project.service.ts
@@ -1,20 +1,17 @@
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Injectable } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { v4 as uuidv4 } from 'uuid';
 
 import { CreateProjectDto } from './dto/CreateProject.dto';
 
-import { Member } from 'src/entity/Member';
-import { Project } from 'src/entity/Project';
+import { Project } from '@entity/Project';
 
 @Injectable()
 export class ProjectService {
   constructor(
     @InjectRepository(Project)
     private readonly projectRepository: Repository<Project>,
-    @InjectRepository(Member)
-    private readonly memberRepository: Repository<Member>,
   ) {}
 
   async create(projectData: CreateProjectDto) {
@@ -31,7 +28,9 @@ export class ProjectService {
       });
 
       return await this.projectRepository.save(newProject);
-    } catch (error) {}
+    } catch (error) {
+      throw new BadRequestException('Erro ao criar projeto');
+    }
   }
 
   async findMany(): Promise<Project[]> {
@@ -39,54 +38,44 @@ export class ProjectService {
   }
 
   async findById(projectId: string): Promise<Project> {
-    return await this.projectRepository.findOne({ where: { id: projectId } });
+    try {
+      const project = await this.projectRepository.findOne({ where: { id: projectId } });
+      if (!project) {
+        throw new NotFoundException(`Projeto com id ${projectId} não encontrado`);
+      }
+      return project;
+    } catch (error) {
+      throw new NotFoundException();
+    }
   }
 
   // TODO: Create a way to delete images after delete a project
   async deleteById(projectId: string) {
-    const deletedProject = await this.projectRepository.delete({ id: projectId });
-    return deletedProject;
+    const project = await this.projectRepository.findOne({
+      where: { id: projectId },
+    });
+    if (!project) {
+      throw new NotFoundException('Não foi possível encontrar o projeto.');
+    }
+    await this.projectRepository.delete({ id: projectId });
   }
 
-  // async updateProject(projectId: string, payload: Partial<Project>) {
-  //   const project = await this.projectRepository.findOne({ where: { id: projectId } });
+  async updateProject(projectId: string, payload: Partial<Project>) {
+    const project = await this.projectRepository.findOne({
+      where: { id: projectId },
+    });
+    if (!project) {
+      return new NotFoundException();
+    }
 
-  //   const data = payload;
-  // }
+    const updatedProject = {
+      ...payload,
+      id: project.id,
+      updatedAt: new Date().toISOString(),
+    };
 
-  // Depracated
-  // async addProject(memberId: string, data: CreateProjectDto) {
-  //   try {
-  //     const member = await this.memberService.findOne(memberId);
-  //     const newProject: OldProject = {
-  //       _id: uuidv4(),
-  //       projectName: data.projectName,
-  //       projectCover: data.projectCover,
-  //       description: data.description,
-  //       technologies: convertToArray(data.technologies),
-  //       projectUrl: data.projectUrl,
-  //     };
-  //     member.projects.push(newProject);
-  //     await this.memberModel.updateOne({ _id: memberId }, { $set: member });
-  //   } catch {
-  //     throw Error();
-  //   }
-  // }
+    await this.projectRepository.update({ id: project.id }, updatedProject);
 
-  // async removeProject(projectId: string, memberId: string): Promise<boolean> {
-  //   try {
-  //     const member = await this.memberService.findOne(memberId);
-  //     if (!member) {
-  //       new BadRequestException('Usuário não encontrado.');
-  //     }
-  //     const project = member.projects.find((p) => p._id === projectId);
-  //     await deleteFile(project.projectCover);
-  //     member.projects = member.projects.filter((p) => p._id !== projectId);
-
-  //     await this.memberModel.updateOne({ _id: memberId }, { $set: member });
-  //     return true;
-  //   } catch {
-  //     return false;
-  //   }
-  // }
+    return updatedProject as Project;
+  }
 }
